added randomn vector projection to feature tokenization

+++ mcpfn/benchmark/get_openml_errors.py
@@ -28,9 +28,11 @@ if "mcpfn" in imputers:
         device='cuda',
         encoder_path='/root/tabular/mcpfn/src/mcpfn/model/encoder.pth',
         borders_path='/root/tabular/mcpfn/borders.pt',
-        checkpoint_path='/root/checkpoints/mar_mixed/step-35000.ckpt'
+        # checkpoint_path='/mnt/mcpfn_data/checkpoints/mixed/step-445000.ckpt'
+        # checkpoint_path='/root/checkpoints/mar_mixed/step-57000.ckpt'
+        checkpoint_path='/mnt/mcpfn_data/checkpoints/mixed_random/step-99999.ckpt'
     )
-    mcpfn_name = "mcpfn_mar_linear"
+    mcpfn_name = "mcpfn_mixed_random"
 
 if "tabpfn" in imputers:
     tabpfn = TabPFNImputer(device='cuda')
@@ -58,31 +60,37 @@ for name in pbar:
 
         # MCPFN
         if "mcpfn" in imputers:
+            # check if file exists. if not, run imputation
+            # if not os.path.exists(f"{base_path}/{name}/{config}/{mcpfn_name}.npy"):
             X_mcpfn = mcpfn.impute(X_missing.copy())
             np.save(f"{base_path}/{name}/{config}/{mcpfn_name}.npy", X_mcpfn)
         
         # TabPFN
         if "tabpfn" in imputers:
-            X_tabpfn = tabpfn.impute(X_missing.copy())
-            np.save(f"{base_path}/{name}/{config}/tabpfn.npy", X_tabpfn)
+            if not os.path.exists(f"{base_path}/{name}/{config}/tabpfn.npy"):
+                X_tabpfn = tabpfn.impute(X_missing.copy())
+                np.save(f"{base_path}/{name}/{config}/tabpfn.npy", X_tabpfn)
         
         # SoftImpute
         if "softimpute" in imputers:
-            plugin = Imputers().get("softimpute")
-            out = plugin.fit_transform(X_missing.copy()).to_numpy()
-            np.save(f"{base_path}/{name}/{config}/softimpute.npy", out)
+            if not os.path.exists(f"{base_path}/{name}/{config}/softimpute.npy"):
+                plugin = Imputers().get("softimpute")
+                out = plugin.fit_transform(X_missing.copy()).to_numpy()
+                np.save(f"{base_path}/{name}/{config}/softimpute.npy", out)
         
         # Column Mean
         if "column_mean" in imputers:
-            plugin = Imputers().get("mean")
-            out = plugin.fit_transform(X_missing.copy()).to_numpy()
-            np.save(f"{base_path}/{name}/{config}/column_mean.npy", out)
+            if not os.path.exists(f"{base_path}/{name}/{config}/column_mean.npy"):
+                plugin = Imputers().get("mean")
+                out = plugin.fit_transform(X_missing.copy()).to_numpy()
+                np.save(f"{base_path}/{name}/{config}/column_mean.npy", out)
         
         # HyperImpute
         if "hyperimpute" in imputers:
-            plugin = Imputers().get("hyperimpute")
-            out = plugin.fit_transform(X_missing.copy()).to_numpy()
-            np.save(f"{base_path}/{name}/{config}/hyperimpute.npy", out)
+            if not os.path.exists(f"{base_path}/{name}/{config}/hyperimpute.npy"):
+                plugin = Imputers().get("hyperimpute")
+                out = plugin.fit_transform(X_missing.copy()).to_numpy()
+                np.save(f"{base_path}/{name}/{config}/hyperimpute.npy", out)
         
         # Empty cache
         torch.cuda.empty_cache()
